requirements.txt
flask==3.0.2
flask-cors==4.0.0
asgiref==3.7.2
SpeechRecognition==3.10.1
pydub==0.25.1
google-generativeai==0.3.2
googletrans==3.1.0a0
pyodbc==5.0.1